containers: Upgrade unit-tests to Ubuntu 18.04 LTS
This tests on a recent toolchain and pyflakes.  Stop installing and
using the "g" npm module, the distro's nodejs is now recent enough.

The main repo -dbg packages are mostly gone now, so enable the dbgsym
repository (as per https://wiki.ubuntu.com/DebuggingProgramCrash) and
install the corresponding -dbgsym packages instead.

Use eatmydata to install the bulk of packages for faster container
creation.

Note that we still test building/running on 16.04 LTS on the ubuntu-1604
image.